# docs/admin/admin_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "接口",
        "title": "管理系统API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/admin/audit_conn/batchDelete": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "链接日志批量删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "链接日志"
                ],
                "summary": "链接日志批量删除",
                "parameters": [
                    {
                        "description": "链接日志",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.AuditConnLogIds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/audit_conn/delete": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "链接日志删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "链接日志"
                ],
                "summary": "链接日志删除",
                "parameters": [
                    {
                        "description": "链接日志信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuditConn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/audit_conn/list": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "链接日志列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "链接日志"
                ],
                "summary": "链接日志列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "目标设备",
                        "name": "peer_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "来源设备",
                        "name": "from_peer",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AuditConnList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/audit_file/batchDelete": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "文件日志批量删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件日志"
                ],
                "summary": "文件日志批量删除",
                "parameters": [
                    {
                        "description": "文件日志",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.AuditFileLogIds"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/audit_file/delete": {
            "post": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "文件日志删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件日志"
                ],
                "summary": "文件日志删除",
                "parameters": [
                    {
                        "description": "文件日志信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AuditFile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/audit_file/list": {
            "get": {
                "security": [
                    {
                        "token": []
                    }
                ],
                "description": "文件日志列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件日志"
                ],
                "summary": "文件日志列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "目标设备",
                        "name": "peer_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "来源设备",
                        "name": "from_peer",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.AuditFileList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.AuditConnLogIds": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "maxItems": 1000,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "admin.AuditFileLogIds": {
            "type": "object",
            "required": [
                "ids"
            ],
            "properties": {
                "ids": {
                    "type": "array",
                    "maxItems": 1000,
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "model.AuditConn": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "client": {
                    "type": "string"
                },
                "close_time": {
                    "type": "integer"
                },
                "conn_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "from_name": {
                    "type": "string"
                },
                "from_peer": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "peer_id": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.AuditConnList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AuditConn"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.AuditFile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "from_name": {
                    "type": "string"
                },
                "from_peer": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "info": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "is_file": {
                    "type": "boolean"
                },
                "num": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "peer_id": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "model.AuditFileList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AuditFile"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "token": {
            "type": "apiKey",
            "name": "api-token",
            "in": "header"
        }
    }
}